feat: allow direct resolve #584
BREAKING CHANGE: direct resolve was denied because it was problematic
together with nested injectors support. Nested injectors feature was
removed, so we don't have a reason to hold to direct resolve
restriction as well.
