+++ src/_dependencies/objects/classes.py
@@ -10,14 +10,12 @@ def _is_class(name, dependency):
 
 def _build_class_spec(name, dependency):
     if _using_object_init(dependency):
-        return _Spec(_ClassFactory(dependency), {}, set(), set(), lambda: None, False)
+        return _Spec(_ClassFactory(dependency), {}, set(), set(), False)
     else:
         name = dependency.__name__ + "." + "__init__"
         owner = f"{dependency.__name__!r} class"
         args, required, optional = _method_args(dependency.__init__, name, owner)
-        return _Spec(
-            _ClassFactory(dependency), args, required, optional, lambda: None, False
-        )
+        return _Spec(_ClassFactory(dependency), args, required, optional, False)
 
 
 def _using_object_init(cls):
