+++ src/_dependencies/spec.py
@@ -1,23 +1,13 @@
-from _dependencies.exceptions import DependencyError
-
-
 class _Spec:
-    def __init__(self, factory, args, required, optional, resolve, is_context):
+    def __init__(self, factory, args, required, optional, is_context):
         _validate_factory(factory)
         _validate_args(args, required, optional)
-        _validate_resolve(resolve)
+        _validate_context(is_context)
         self.factory = factory
         self.args = args
         self.required = required
         self.optional = optional
-        self.resolve = resolve
         self.is_context = is_context
-
-    def resolved(self):
-        kind = self.resolve()
-        if kind is not None:
-            message = f"{kind} dependencies could only be used to instantiate classes"
-            raise DependencyError(message)
 
 
 def _validate_factory(factory):
@@ -30,6 +20,6 @@ def _validate_args(args, required, optional):
         raise RuntimeError
 
 
-def _validate_resolve(resolve):
-    if not callable(resolve):
+def _validate_context(is_context):
+    if not isinstance(is_context, bool):
         raise RuntimeError
