--- tests/test_direct_resolve.py
@@ -1,37 +1,0 @@
-"""Tests related to direct resolve rules."""
-from dependencies import Injector
-from dependencies import value
-
-
-def test_direct_resolve_scalar(expect, catch):
-    """Attempt to resolve scalar types directly should raise exception.
-
-    Scalar types are allowed to be used as dependencies for classes.
-
-    """
-
-    class Container(Injector):
-        a = 1
-
-    @expect(Container)
-    @catch("Scalar dependencies could only be used to instantiate classes")
-    def case(it):
-        it.a
-
-
-def test_direct_resolve_value(expect, catch):
-    """Attempt to resolve value directly should raise exception.
-
-    Values are allowed to be used as dependencies for classes.
-
-    """
-
-    class Container(Injector):
-        @value
-        def a():
-            return 1
-
-    @expect(Container)
-    @catch("'value' dependencies could only be used to instantiate classes")
-    def case(it):
-        it.a
